Revert "added more handlers"
This reverts commit 23b98b7f63e27785022a3a498be8ac8d57616399.
